Extract user from pyconkr
