+++ pyconkr/admin.py
@@ -10,14 +10,15 @@ from modeltranslation.admin import TranslationAdmin
 from sorl.thumbnail.admin import AdminImageMixin
 from .models import (Room, Program, ProgramTime, ProgramDate, ProgramCategory,
                      Speaker, Sponsor, SponsorLevel, Preference,
-                     Profile, Announcement, EmailToken, Proposal, Banner,
+                     Announcement, EmailToken, Proposal, Banner,
                      TutorialProposal, TutorialCheckin, SprintProposal, SprintCheckin)
 from .actions import convert_proposal_to_program
 
 
 class SummernoteWidgetWithCustomToolbar(SummernoteWidget):
     def template_contexts(self):
-        contexts = super(SummernoteWidgetWithCustomToolbar, self).template_contexts()
+        contexts = super(SummernoteWidgetWithCustomToolbar,
+                         self).template_contexts()
         contexts['width'] = '960px'
         return contexts
 
@@ -26,11 +27,15 @@ class RoomAdmin(SummernoteModelAdmin, TranslationAdmin):
     list_display = ('id', 'name',)
     list_editable = ('name',)
     search_fields = ('name',)
+
+
 admin.site.register(Room, RoomAdmin)
 
 
 class ProgramDateAdmin(admin.ModelAdmin):
     list_display = ('id', 'day',)
+
+
 admin.site.register(ProgramDate, ProgramDateAdmin)
 
 
@@ -38,21 +43,28 @@ class ProgramTimeAdmin(TranslationAdmin):
     list_display = ('id', 'name', 'begin', 'end', 'day')
     list_editable = ('name', 'day')
     ordering = ('begin',)
+
+
 admin.site.register(ProgramTime, ProgramTimeAdmin)
 
 
 class ProgramCategoryAdmin(TranslationAdmin):
     list_display = ('id', 'name', 'slug',)
     list_editable = ('name', 'slug',)
+
+
 admin.site.register(ProgramCategory, ProgramCategoryAdmin)
 
 
 class SponsorAdmin(SummernoteModelAdmin, TranslationAdmin):
-    formfield_overrides = {models.TextField: {'widget': SummernoteWidgetWithCustomToolbar}}
+    formfield_overrides = {models.TextField: {
+        'widget': SummernoteWidgetWithCustomToolbar}}
     list_display = ('id', 'slug', 'name',)
     ordering = ('name',)
     list_editable = ('slug', 'name',)
     search_fields = ('name',)
+
+
 admin.site.register(Sponsor, SponsorAdmin)
 
 
@@ -61,6 +73,8 @@ class SponsorLevelAdmin(SummernoteModelAdmin, TranslationAdmin):
     list_editable = ('order', 'name', 'slug',)
     ordering = ('order',)
     search_fields = ('name',)
+
+
 admin.site.register(SponsorLevel, SponsorLevelAdmin)
 
 
@@ -69,15 +83,20 @@ class SpeakerAdmin(SummernoteModelAdmin, TranslationAdmin):
     list_editable = ('slug', 'name', 'email',)
     ordering = ('name',)
     search_fields = ('name', 'slug', 'email',)
+
+
 admin.site.register(Speaker, SpeakerAdmin)
 
 
 class ProgramAdmin(SummernoteModelAdmin, TranslationAdmin):
-    list_display = ('id', 'name', 'date', 'room', 'slide_url', 'pdf_url', 'get_speakers', 'category', 'is_recordable',)
+    list_display = ('id', 'name', 'date', 'room', 'slide_url',
+                    'pdf_url', 'get_speakers', 'category', 'is_recordable',)
     list_editable = ('name', 'category', 'is_recordable',)
     ordering = ('id',)
     filter_horizontal = ('times', )
     search_fields = ('name', 'speakers__name', 'desc',)
+
+
 admin.site.register(Program, ProgramAdmin)
 
 
@@ -85,26 +104,26 @@ class AnnouncementAdmin(SummernoteModelAdmin, TranslationAdmin):
     list_display = ('id', 'title', 'created', 'modified')
     ordering = ('id',)
     search_fields = ('title',)
+
+
 admin.site.register(Announcement, AnnouncementAdmin)
 
 
 class EmailTokenAdmin(admin.ModelAdmin):
     list_display = ('email', 'token', 'created')
     search_fields = ('email',)
+
+
 admin.site.register(EmailToken, EmailTokenAdmin)
 
 
 class FlatPageAdmin(TranslationAdmin):
-    formfield_overrides = {models.TextField: {'widget': SummernoteWidgetWithCustomToolbar}}
+    formfield_overrides = {models.TextField: {
+        'widget': SummernoteWidgetWithCustomToolbar}}
 
 
 admin.site.unregister(FlatPage)
 admin.site.register(FlatPage, FlatPageAdmin)
-
-
-class ProfileInline(AdminImageMixin, admin.StackedInline):
-    model = Profile
-    can_delete = False
 
 
 class ProposalAdminForm(forms.ModelForm):
@@ -121,6 +140,8 @@ class ProposalAdmin(admin.ModelAdmin):
     form = ProposalAdminForm
     list_display = ('user', 'title', 'difficulty', 'duration', 'language')
     actions = [convert_proposal_to_program]
+
+
 admin.site.register(Proposal, ProposalAdmin)
 
 
@@ -139,6 +160,8 @@ class TutorialProposalAdmin(admin.ModelAdmin):
     list_display = ('user', 'title', 'difficulty', 'duration', 'language', 'capacity',
                     'begin_date', 'begin_time', 'end_date', 'end_time', )
     actions = [convert_proposal_to_program]
+
+
 admin.site.register(TutorialProposal, TutorialProposalAdmin)
 
 
@@ -154,34 +177,39 @@ class SprintProposalAdminForm(forms.ModelForm):
 
 class SprintProposalAdmin(admin.ModelAdmin):
     form = SprintProposalAdminForm
-    list_display = ('title', 'language', 'project_url', 'project_brief', 'contribution_desc')
+    list_display = ('title', 'language', 'project_url',
+                    'project_brief', 'contribution_desc')
     actions = [convert_proposal_to_program]
+
+
 admin.site.register(SprintProposal, SprintProposalAdmin)
-
-
-class UserAdmin(BaseUserAdmin):
-    inlines = (ProfileInline, )
-
-admin.site.unregister(User)
-admin.site.register(User, UserAdmin)
 
 
 class BannerAdmin(SummernoteModelAdmin, TranslationAdmin):
     list_display = ('id', 'name', 'url', 'begin', 'end')
     ordering = ('id',)
     search_fields = ('name', 'url')
+
+
 admin.site.register(Banner, BannerAdmin)
 
 
 class PreferenceAdmin(admin.ModelAdmin):
     list_display = ('id', 'user', 'program',)
+
+
 admin.site.register(Preference, PreferenceAdmin)
 
 
 class TutorialCheckinAdmin(admin.ModelAdmin):
     list_display = ('id', 'user', 'tutorial',)
+
+
 admin.site.register(TutorialCheckin, TutorialCheckinAdmin)
+
 
 class SprintCheckinAdmin(admin.ModelAdmin):
     list_display = ('id', 'user', 'sprint',)
+
+
 admin.site.register(SprintCheckin, SprintCheckinAdmin)
