+++ user/admin.py
@@ -1,0 +1,20 @@
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
+from django.contrib.auth import get_user_model
+from sorl.thumbnail.admin import AdminImageMixin
+from .models import Profile
+
+User = get_user_model()
+
+
+class ProfileInline(AdminImageMixin, admin.StackedInline):
+    model = Profile
+    can_delete = False
+
+
+class UserAdmin(BaseUserAdmin):
+    inlines = (ProfileInline, )
+
+
+admin.site.unregister(User)
+admin.site.register(User, UserAdmin)
